+++ superduperdb/backends/mongodb/artifacts.py
@@ -1,3 +1,7 @@
+import os
+import tempfile
+from pathlib import Path
+
 import click
 import gridfs
 
@@ -48,6 +52,28 @@ class MongoArtifactStore(ArtifactStore):
             raise FileNotFoundError(f'File not found in {file_id}')
         return cur.read()
 
+    def _save_file(self, file_path: str, file_id: str):
+        path = Path(file_path)
+        if path.is_dir():
+            upload_folder(file_path, file_id, self.filesystem)
+        else:
+            self.filesystem.put(
+                open(file_path, 'rb'),
+                filename=file_path,
+                metadata={"file_id": file_id, "type": "file"},
+            )
+
+    def _load_file(self, file_id: str) -> str:
+        file = self.filesystem.find_one(
+            {"metadata.file_id": file_id, "metadata.type": "file"}
+        )
+        if file is not None:
+            with open(file.filename, 'wb') as f:
+                f.write(file.read())
+            return file.filename
+
+        return download_folder(file_id, self.filesystem)
+
     def _save_bytes(self, serialized: bytes, file_id: str):
         return self.filesystem.put(serialized, filename=file_id)
 
@@ -57,3 +83,43 @@ class MongoArtifactStore(ArtifactStore):
         """
 
         # TODO: implement me
+
+
+def upload_folder(path, file_id, fs, parent_path=""):
+    if not os.listdir(path):
+        fs.put(
+            b'',
+            filename=os.path.join(parent_path, os.path.basename(path) + '/'),
+            metadata={"file_id": file_id, "is_empty_dir": True, 'type': 'dir'},
+        )
+    else:
+        for item in os.listdir(path):
+            item_path = os.path.join(path, item)
+            if os.path.isdir(item_path):
+                upload_folder(item_path, file_id, fs, os.path.join(parent_path, item))
+            else:
+                with open(item_path, 'rb') as file_to_upload:
+                    fs.put(
+                        file_to_upload,
+                        filename=os.path.join(parent_path, item),
+                        metadata={"file_id": file_id, "type": "dir"},
+                    )
+
+
+def download_folder(file_id, fs):
+    temp_dir = tempfile.mkdtemp()
+    logging.info(f"Downloading files to temporary directory: {temp_dir}")
+
+    for grid_out in fs.find({"metadata.file_id": file_id, "metadata.type": "dir"}):
+        file_path = os.path.join(temp_dir, grid_out.filename)
+        if grid_out.metadata.get("is_empty_dir", False):
+            if not os.path.exists(file_path):
+                os.makedirs(file_path)
+        else:
+            directory = os.path.dirname(file_path)
+            if not os.path.exists(directory):
+                os.makedirs(directory)
+            with open(file_path, 'wb') as file_to_write:
+                file_to_write.write(grid_out.read())
+
+    return temp_dir